# [HotFix] Stop tracking syncing live WAL for performance