--- a/tools/db_bench_tool.cc
+++ b/tools/db_bench_tool.cc
@@ -1681,6 +1681,9 @@ DEFINE_uint32(write_batch_protection_bytes_per_key, 0,
 DEFINE_bool(build_info, false,
             "Print the build info via GetRocksBuildInfoAsString");
 
+DEFINE_bool(track_and_verify_wals_in_manifest, false,
+            "If true, enable WAL tracking in the MANIFEST");
+
 namespace ROCKSDB_NAMESPACE {
 namespace {
 static Status CreateMemTableRepFactory(
@@ -4485,6 +4488,8 @@ class Benchmark {
     }
 
     options.allow_data_in_errors = FLAGS_allow_data_in_errors;
+    options.track_and_verify_wals_in_manifest =
+        FLAGS_track_and_verify_wals_in_manifest;
 
     // Integrated BlobDB
     options.enable_blob_files = FLAGS_enable_blob_files;