@riversand963 2022-07-11
Thanks a lot for the insightful feedback @ajkr!

I should have mentioned that this PR is a hot-fix that restores to the original behavior. The purpose is to unblock MySQL/MyRocks.
In addition to this PR (after it's merged), I plan to extend track_and_verify_wals_in_manifest to be more flexible. Applications can choose whether to track sync operations on the live WAL or not. The flexibility is helpful because some applications need RocksDB to fail recovery when detecting loss of synced data in the WAL, while other applications allow RocksDB to succeed in recovery with loss of synced data, as long as the application knows which data is lost. MyRocks falls into the latter.

> MyRocks team said the impact to them is MANIFEST becomes larger. They can reduce the manifest rollover size if it's a problem.

There is more to this. Having MANIFEST files rotate faster still writes more data to flash, which is undesirable.

With the extension to `track_and_verify_wals`, applications using `rocksdb_flush_log_at_trx_commit ` can benefit from this feature, too, as long as they can use replication to tolerate synced data loss, as in the case of MyRocks.