@riversand963 2022-07-12
@ajkr 
Agree. It is case by case, and we should make it explicit in the API comment. Will update later.