@ajkr 2022-07-12
I think for now it can be as simple as saying only closed WALs are tracked. That is a big limitation that is not obvious in the option name or API doc.