@riversand963 2022-07-12
hmm, since #10087 is already in 7.4 release, I should update HISTORY.md to mention this behavior change as well.